make kubeapply a function
